Restore go1.13 compatibility, but with go.mod specifying go1.17 for cleaner imports.
