+++ .travis.yml
@@ -1,6 +1,9 @@
 language: go
 
 go:
+  - 1.13
+  - 1.14
+  - 1.15
   - 1.16
   - 1.17
   - tip
